# src/main/java/org/openapitools/api/ScheduleApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (6.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package org.openapitools.api;

import org.openapitools.model.ErrorData;
import java.util.List;
import org.openapitools.model.ScheduleDateSlot;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import javax.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2022-11-02T19:55:19.584688500+03:00[Europe/Moscow]")
@Validated
@Tag(name = "Schedule", description = "the schedule API")
@RequestMapping("${openapi.open.base-path:/api/v1}")
public interface ScheduleApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /schedule : Метод добавления расписания работы мобильного банкира
     *
     * @param scheduleDateSlot График работы за несколько дней. Не менее одного дня (required)
     * @return Успешное выполнение (status code 200)
     *         or Ошибка (status code 400)
     */
    @Operation(
        operationId = "addMobileBankerSchedule",
        summary = "Метод добавления расписания работы мобильного банкира",
        tags = { "Schedule" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Успешное выполнение"),
            @ApiResponse(responseCode = "400", description = "Ошибка", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorData.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/schedule",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<Void> addMobileBankerSchedule(
        @Parameter(name = "ScheduleDateSlot", description = "График работы за несколько дней. Не менее одного дня", required = true) @Valid @RequestBody List<ScheduleDateSlot> scheduleDateSlot
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    /**
     * GET /schedule : Метод получения расписания мобильного банкира
     *
     * @param startDate Начало диапазона графика работы (required)
     * @param endDate Конец диапазона графика работы (required)
     * @return Успешное выполнение (status code 200)
     *         or Ошибка (status code 400)
     */
    @Operation(
        operationId = "getMobileBankerSchedule",
        summary = "Метод получения расписания мобильного банкира",
        tags = { "Schedule" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Успешное выполнение", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ScheduleDateSlot.class))
            }),
            @ApiResponse(responseCode = "400", description = "Ошибка", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorData.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/schedule",
        produces = { "application/json" }
    )
    default ResponseEntity<List<ScheduleDateSlot>> getMobileBankerSchedule(
        @NotNull @Parameter(name = "startDate", description = "Начало диапазона графика работы", required = true) @Valid @RequestParam(value = "startDate", required = true) String startDate,
        @NotNull @Parameter(name = "endDate", description = "Конец диапазона графика работы", required = true) @Valid @RequestParam(value = "endDate", required = true) String endDate
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"scheduleDate\" : \"02.10.2022\", \"isHoliday\" : false, \"timeSlots\" : [ { \"slotEndTime\" : \"780\", \"slotStartTime\" : \"720\", \"slotId\" : 2, \"isWorking\" : true }, { \"slotEndTime\" : \"780\", \"slotStartTime\" : \"720\", \"slotId\" : 2, \"isWorking\" : true } ], \"isWeekend\" : true }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}