# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Мобильный банкир OpenAPi
  version: 0.0.1
servers:
- description: Local stend
  url: http://localhost:8080/api/v1
paths:
  /mobile-banker:
    get:
      operationId: getMobileBanker
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MobileBanker'
          description: Успешное получение - Профиль МБ
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorData'
          description: Ошибка
      summary: Метод получения мобильного банкира
      tags:
      - Mobile banker
      x-accepts: application/json
      x-tags:
      - tag: Mobile banker
  /mobile-banker/{employeeId}/deleting-device-id:
    put:
      operationId: deleteDeviceId
      parameters:
      - description: ID сотрудника
        example: 123456-7890-abcd-ef0123
        explode: false
        in: path
        name: employeeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json: {}
          description: Успешное выполнение
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorData'
          description: Ошибка
      summary: Метод удаления device-id мобильного банкира
      tags:
      - Mobile banker
      x-accepts: application/json
      x-tags:
      - tag: Mobile banker
  /schedule:
    get:
      operationId: getMobileBankerSchedule
      parameters:
      - description: Начало диапазона графика работы
        example: 01.11.2022
        explode: true
        in: query
        name: startDate
        required: true
        schema:
          type: string
        style: form
      - description: Конец диапазона графика работы
        example: 02.11.202
        explode: true
        in: query
        name: endDate
        required: true
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
          description: Успешное выполнение
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorData'
          description: Ошибка
      summary: Метод получения расписания мобильного банкира
      tags:
      - Schedule
      x-accepts: application/json
      x-tags:
      - tag: Schedule
    post:
      operationId: addMobileBankerSchedule
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
        description: График работы за несколько дней. Не менее одного дня
        required: true
      responses:
        "200":
          content:
            application/json: {}
          description: Успешное выполнение
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorData'
          description: Ошибка
      summary: Метод добавления расписания работы мобильного банкира
      tags:
      - Schedule
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Schedule
...
components:
...
  schemas:
    MobileBanker:
      example:
        employeeFullName: Иванов Иван Иванович
        phone: "74956735665"
        city: Москва
        photo: photo
        position: Менеджер
        supervisorFullName: Cидоров Василий Иванович
        email: example@mail.com
      properties:
        employeeFullName:
          description: Полное ФИО МБ
          example: Иванов Иван Иванович
          maxLength: 150
          type: string
        photo:
          description: Ссылка на фото сотрудника на ресурсе Банка.
          maxLength: 1024
          type: string
        email:
          description: Электронная почта
          example: example@mail.com
          maxLength: 50
          type: string
        phone:
          description: Мобильный телефон
          example: "74956735665"
          maxLength: 15
          pattern: "^(\\+)\\d{11}$"
          type: string
        position:
          description: специалист
          example: Менеджер
          maxLength: 50
          type: string
        city:
          description: Город
          example: Москва
          maxLength: 50
          type: string
        supervisorFullName:
          description: Анна Андреевна Семёновна
          example: Cидоров Василий Иванович
          maxLength: 150
          type: string
      required:
      - city
      - email
      - employeeFullName
      - phone
      - position
      - supervisorFullName
      type: object
...
    TimeSlot:
      example:
        slotEndTime: "780"
        slotStartTime: "720"
        slotId: 2
        isWorking: true
      properties:
        slotId:
          description: Порядковый номер слота
          example: 2
          type: integer
        slotStartTime:
          description: Время начала слота
          example: "720"
          format: 00:00
          type: string
        slotEndTime:
          description: Время окончания слота
          example: "780"
          format: 00:00
          type: string
        isWorking:
          description: Признак рабочего слота
          example: true
          type: boolean
      required:
      - isWorking
      - slotEndTime
      - slotId
      - slotStartTime
      type: object
    ScheduleDateSlot:
      example:
        scheduleDate: 02.10.2022
        isHoliday: false
        timeSlots:
        - slotEndTime: "780"
          slotStartTime: "720"
          slotId: 2
          isWorking: true
        - slotEndTime: "780"
          slotStartTime: "720"
          slotId: 2
          isWorking: true
        isWeekend: true
      properties:
        scheduleDate:
          description: День в графике
          example: 02.10.2022
          format: 27.10.2022
          type: string
        isHoliday:
          description: "Признак нерабочего дня (праздник, распоряжение по компании\
            \ и пр.)"
          example: false
          type: boolean
        timeSlots:
          items:
            $ref: '#/components/schemas/TimeSlot'
          type: array
        isWeekend:
          description: Признак выходного дня
          example: true
          type: boolean
      required:
      - isHoliday
      - isWeekend
      - scheduleDate
      - timeSlots
      type: object
    Schedule:
      items:
        $ref: '#/components/schemas/ScheduleDateSlot'
      type: array
...
    ErrorData:
      description: Данные об ошибке
      properties:
        errorCode:
          description: Код ошибки
          example: 12
          type: integer
        errorMessage:
          description: Описание ошибки
          example: Не удалось загрузить файл
          maxLength: 128
          type: string
      required:
      - errorCode
      - errorMessage
      type: object
...